Updated branching, and moved to seperate file.
